use authenticated supabase client in api routes

+++ src/app/api/data/DELETE.ts
@@ -1,18 +1,22 @@
-import * as z from "zod";
-import { getSupabaseClient } from "@/lib/supabase";
+import { createServerClient } from "@/lib/supabase/createServerClient"
 import { NextRequest } from "next/server";
 
-export async function DELETE(
-  request: Request) {
+export async function DELETE(request: NextRequest) {
   const { searchParams } = new URL(request.url)
   const slug = searchParams.get('slug')
-  const supabase = getSupabaseClient();
+  const supabase = await createServerClient()
+  const user = (await supabase.auth.getUser()).data?.user
+
+  if(!user) {
+    return Response.json({ error: "Unauthorized" }, { status: 401 })
+  }
 
   if (slug) {
     const { error } = await supabase
       .from('datasets')
       .delete()
-      .eq('dataset_slug', slug);
+      .eq('dataset_slug', slug)
+      // .eq('owner_id', user.id)
 
     if (error) {
       return Response.json({ error: "Failed to delete dataset" }, { status: 500 });
